requirements.txt
# Core dependencies
streamlit==1.40.0
youtube-transcript-api>=0.6.1
langchain-google-genai>=2.0.0
langchain-core>=0.3.15
google-generativeai>=0.8.0
langdetect==1.0.9

# SSL Certificate handling
certifi>=2024.2.2
urllib3>=2.0.7,<3.0

# Utilities
python-dotenv==1.0.1
tqdm==4.67.1
rich==13.9.4

# Visualization
plotly>=5.18.0

# ML/Embeddings
sentence-transformers>=2.2.0
torch>=2.0.0
transformers>=4.30.0

# Clustering
hdbscan>=0.8.33
umap-learn>=0.5.4

# YouTube API
google-api-python-client>=2.100.0